+++ src/middleware/order.js
@@ -1,6 +1,6 @@
 import rejectingFetch from './rejectingFetch';
 import invariant from 'invariant';
-import { dataApiPath } from './paths';
+import { orderApiPath } from './paths';
 import { headersGet, headersPost, headersPut, headersDelete } from './headers';
 import Order from '../models/Order';
 
@@ -11,7 +11,7 @@ export const submitOrder = (order, foundry = 'egf') => {
   invariant(foundry === 'egf', 'must submit a foundry (right now, only egf works');
   invariant(Order.validateSetup(order), 'order be valid partial order (prior to ID + foundry data)');
 
-  const url = dataApiPath(`order/${order.projectId}`);
+  const url = orderApiPath(`${order.projectId}`);
   const stringified = JSON.stringify({
     foundry,
     order,
@@ -26,14 +26,14 @@ const getQuote = (foundry, order) => {
 };
 
 export const getOrder = (projectId, orderId) => {
-  const url = dataApiPath(`order/${projectId}/${orderId}`);
+  const url = orderApiPath(`${projectId}/${orderId}`);
 
   return rejectingFetch(url, headersGet())
     .then(resp => resp.json());
 };
 
 export const getOrders = (projectId) => {
-  const url = dataApiPath(`order/${projectId}`);
+  const url = orderApiPath(`${projectId}`);
 
   return rejectingFetch(url, headersGet())
     .then(resp => resp.json());
