+++ src/middleware/paths.js
@@ -6,6 +6,7 @@ export const serverRoot = (global.location && (/http/gi).test(global.location.pr
 
 export const authPath = (path) => serverRoot + 'auth/' + path;
 export const dataApiPath = (path) => serverRoot + 'data/' + path;
+export const orderApiPath = (path) => serverRoot + 'order/' + path;
 export const extensionsPath = (id) => serverRoot + 'extensions/' + id;
 export const fileApiPath = (path) => serverRoot + 'file/' + path;
 export const importPath = (path) => serverRoot + 'import/' + path;
